Revert part of revision 166
The piece of code that was removed was for distinguishing between "indicator is installed" and "indicator is installed AND active/running".
Removing it caused specto to not quit when closing the window and indicator applet is not running, preventing the user from bring the window back except by toggling a key in gconf.